% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drive_time_polygons.R
\name{get_drivetime_polygons}
\alias{get_drivetime_polygons}
\title{Get drive time polygons around points}
\usage{
get_drivetime_polygons(lat, lon, drive_time, res = 30)
}
\arguments{
\item{lat}{Vector of latitudes}

\item{lon}{Vector of longitudes}

\item{drive_time}{Length of drive time in minutes (e.g. 30 for 30 minutes)}

\item{res}{Resolution. Higher resolution yields more detailed polygons and takes more time.}
}
\value{
A vector of drive time polygons
}
\description{
Get drive time polygons around points
}
