% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_miles.R
\name{road_miles}
\alias{road_miles}
\title{Calculate total road miles in a county}
\usage{
road_miles(county_geoid, year = 2019, output_unit = c("miles", "km", "meters"))
}
\arguments{
\item{county_geoid}{5 character county GEOID}

\item{year}{Year of the data to download}

\item{output_unit}{Units for the output}
}
\value{
A numeric vector representing total roadmiles in the county
}
\description{
Calculate total road miles in a county
}
